refactor: extract test_fixtures crate from integration_tests
