+++ test_fixtures/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "e2e_bench"
+name = "test_fixtures"
 version = "0.1.0"
 edition = "2024"
 license = { workspace = true }
@@ -11,19 +11,24 @@ workspace = true
 [dependencies]
 common.workspace = true
 indexer_service.workspace = true
-indexer_service_rpc = { workspace = true, features = ["client"] }
+key_protocol.workspace = true
 nssa.workspace = true
+nssa_core = { workspace = true, features = ["host"] }
+sequencer_core = { workspace = true, features = ["default", "testnet"] }
 sequencer_service.workspace = true
 sequencer_service_rpc = { workspace = true, features = ["client"] }
-test_fixtures.workspace = true
+vault_core.workspace = true
 wallet.workspace = true
 
 anyhow.workspace = true
-borsh.workspace = true
-chrono.workspace = true
-clap.workspace = true
+bytesize.workspace = true
+env_logger.workspace = true
+futures.workspace = true
 jsonrpsee = { workspace = true, features = ["ws-client"] }
+log.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tempfile.workspace = true
-tokio = { workspace = true, features = ["macros", "rt-multi-thread", "time"] }
+testcontainers = { version = "0.27.3", features = ["docker-compose"] }
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
+url.workspace = true
