+++ test_fixtures/src/indexer_client.rs
