+++ tools/e2e_bench/src/scenarios/fanout.rs
@@ -3,7 +3,7 @@
 use std::time::Instant;
 
 use anyhow::{Result, bail};
-use integration_tests::public_mention;
+use test_fixtures::public_mention;
 use wallet::cli::{
     Command, SubcommandReturnValue,
     account::{AccountSubcommand, NewSubcommand},
